Change type of port parameters to uint16_t
This causes Python to throw OverflowError on bad values, rather than
propagating them into C++ where they will wrap around.

Also added tests for bad port numbers.

Closes #8.

+++ src/recv_netmap.h
@@ -73,7 +73,7 @@ public:
      * @param device       Name of the network interface e.g., @c eth0
      * @param port         UDP port number to listen to
      */
-    netmap_udp_reader(stream &owner, const std::string &device, int port);
+    netmap_udp_reader(stream &owner, const std::string &device, uint16_t port);
 
     virtual void stop() override;
 };
